Added Testing Maze
I added the testing maze to prefabs incase anybody wants to test out lighting, movement, etc., but DON'T OVERRIDE THE PREFAB. If you do, it may cause a lot of bugs since others may be using it. The maze is generally small, and you can activate a cieling if you're testing a flashlight/torch design. Also added a material for it so that when the light is generated, it doesn't look blinding.